backend/requirements.txt
Flask==3.0.0
gunicorn==22.0.0
flask-cors==4.0.0
tensorflow-cpu==2.15.0  # Version stable
numpy==1.26.4  # Compatible avec tensorflow 2.15
opencv-python-headless==4.9.0.80  # Compatible avec numpy <2
Pillow==10.2.0
pymongo==4.7.2
mongoengine==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.1
python-json-logger==2.0.7
prometheus-client==0.19.0
cryptography==42.0.0
pytest==8.0.0
pytest-cov==4.1.0
black==24.1.1
flake8==7.0.0
mypy==1.8.0
redis==5.0.1